+++ libs/src/ui/element/Checkbox/Checkbox.vue
@@ -7,101 +7,105 @@ const modelValue = defineModel();
 
 // 定義 Props
 const props = defineProps({
-	label: {
-		type: String,
-	},
-	value: {
-		type: [String, Boolean],
-	},
-	name: {
-		type: String,
-	},
-	id: {
-		type: String,
-	},
 	themeColor: {
 		type: String,
 		default: "primary",
 		validator: (value) =>
 			["primary", "secondary", "tertiary", "success", "warning", "error", "info"].includes(value),
 	},
+	dataSource: {
+		type: Array,
+		default: () => [],
+	},
+	initValue: {
+		type: Array,
+		default: () => [],
+	},
+	direction: {
+		type: String,
+		default: "row",
+		validator: (value) =>
+			["row", "column"].includes(value),
+	},
 	className: {
 		type: String,
 		default: "",
 	},
-	isChecked: {
-		type: Boolean,
-		default: false,
-	},
-	isMultiple: {
-		type: Boolean,
-		default: false,
-	},
 });
 
-// 處理 check 狀態顯示
-const isCheck = ref(false);
+// 用於儲存每個選項的選中狀態
+const isCheck = ref([]);
 
-// 根據傳入的 isChecked 和 modelValue 設置初始值
+// 根據傳入的 initValue 和 modelValue 設置初始值
 watch(
-	() => props.isChecked,
+	() => props.initValue,
 	(newVal) => {
-		if (props.isMultiple) {
-			if (newVal && Array.isArray(modelValue.value) && !modelValue.value.includes(props.value)) {
-				modelValue.value.push(props.value);
-			} else if (!newVal && Array.isArray(modelValue.value) && modelValue.value.includes(props.value)) {
-				modelValue.value = modelValue.value.filter((item) => item !== props.value);
-			}
-		} else {
-			modelValue.value = newVal;
-		}
-		isCheck.value = newVal;
+		const checkedValues = newVal;
+		// 初始化選中狀態
+		isCheck.value = props.dataSource.map((item) => checkedValues.includes(item.value));
+		modelValue.value = [...checkedValues];
 	},
-	{ immediate: true }
+	{immediate: true}
 );
 
 // 切換選中狀態
-const handleCheck = () => {
-	if (props.isMultiple) {
-		if (Array.isArray(modelValue.value)) {
-			if (isCheck.value) {
-				modelValue.value = modelValue.value.filter((item) => item !== props.value);
-			} else {
-				modelValue.value.push(props.value);
-			}
+const handleCheck = (item, index) => {
+	const isMultiple = Array.isArray(modelValue.value);
+
+	if (isMultiple) {
+		if (isCheck.value[index]) {
+			modelValue.value = modelValue.value.filter((value) => value !== item.value);
+		} else {
+			modelValue.value.push(item.value);
 		}
 	} else {
-		modelValue.value = !isCheck.value;
+		modelValue.value = !isCheck.value[index];
 	}
-	isCheck.value = !isCheck.value;
+	isCheck.value[index] = !isCheck.value[index];
 };
 </script>
 
 <template>
-	<label :for="props.id" :class="{'checkbox': true, [props.className]: !!props.className}">
-		<input
-			class="checkbox-input"
-			type="checkbox"
-			:id="props.id"
-			:name="props.name"
-			:value="props.value"
-			:checked="isCheck"
-			@change="handleCheck"
-		/>
-		<!-- checkbox - 選擇框樣式 -->
-		<div
-			:class="[
-                'checkbox-icon',
-                isCheck
-                    ? `checkbox-checked-${props.themeColor}`
-                    : `checkbox-unchecked-${props.themeColor}`,
-            ]"
+	<div :class="{
+		'checkbox-container': true,
+		[`checkbox-container-${props.direction}`]: props.direction,
+		[props.className]: !!props.className
+		}">
+		<label
+			v-for="(item, index) in props.dataSource"
+			:key="item.id"
+			:for="item.id"
+			:class="{'checkbox': true}"
 		>
-			<Icon v-if="isCheck" name="check" color="#fff"></Icon>
-		</div>
-		<!-- checkbox - 選項文字 -->
-		<span class="checkbox-text">{{ props.label }}</span>
-	</label>
+			<input
+				class="checkbox-input"
+				type="checkbox"
+				:id="item.id"
+				:name="item.name"
+				:value="item.value"
+				:checked="isCheck[index]"
+				@change="handleCheck(item, index)"
+			/>
+			<!-- checkbox - 選擇框樣式 -->
+			<div
+				:class="[
+					'checkbox-icon',
+					isCheck[index]
+						? `checkbox-checked-${props.themeColor}`
+						: `checkbox-unchecked-${props.themeColor}`,
+				]"
+			>
+				<Icon v-if="isCheck[index]" name="check" color="#fff"></Icon>
+			</div>
+			<!-- checkbox - 選項文字 -->
+			<span class="checkbox-text">{{ item.label }}</span>
+		</label>
+	</div>
 </template>
 
-<style scoped lang="scss"></style>
+<style scoped lang="scss">
+.checkbox-group {
+	display: flex;
+	flex-direction: column;
+}
+</style>