+++ libs/src/ChangeLog.md
@@ -1,4 +1,4 @@
-## V1.0.0 (2024-10-21)
+## V1.0.0 (2024-11-10)
 
 ✨ New Feature
 
@@ -6,7 +6,9 @@
 
 🆙update
 
-- 同步學長版 storybook
+- 修改 checkbox vue 元件寫法
+- 修改 checkbox story code 顯示方式
+- 修改 checkbox story table 顯示方式
 
 ---